Product Life Cycle - move the M_InOut interceptor out of de.metas.inout.model.validator
The new interceptor was added as de.metas.business's
de.metas.inout.model.validator.M_InOut - the exact fully-qualified name
de.metas.swat.base already uses. Both jars are on the runtime classpath and
de.metas.swat.base depends on de.metas.business, so the new class won the
lookup and SwatValidator ended up registering it instead of swat's, silently
dropping that interceptor's callbacks: captured locations + rendered addresses
(InOutLocationsUpdater), BPartner balance add/remove, quality issues on
complete, shipment-statistics cache reset, description/description-bottom, and
the guard against voiding/reversing an invoiced shipment.

Visible effect: material receipts got no rendered address, so the partner name
was missing from the receipt PDF and the receipt.spec.js / partial-receipt.spec.js
PDF assertions failed in both languages.

Moved to de.metas.product.model.interceptor.M_InOut (next to the M_Product
interceptor), which keeps the @Interceptor + @Component auto-registration and
leaves swat's FQN untouched. Test moved along; no behaviour change.
