+++ backend/de.metas.business/src/test/java/de/metas/product/model/interceptor/M_InOutTest.java
@@ -1,4 +1,4 @@
-package de.metas.inout.model.validator;
+package de.metas.product.model.interceptor;
 
 /*
  * #%L
